% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSelenium.R
\name{start_remote_driver}
\alias{start_remote_driver}
\title{Create browser connection for RSelenium}
\usage{
start_remote_driver(...)
}
\arguments{
\item{...}{\code{ANY}\cr
Parameters passed to \code{\link[RSelenium:remoteDriver-class]{RSelenium::remoteDriver()}}.}
}
\value{
Nothing, function called for its side effects.
}
\description{
Connects to a browser for RSelenium. RSelenium is needed for some cases where
rvest package doesn't suffice, usually due to 403 or 503 status codes
received. However, RSelenium setup is much more complicated, this is why
rvest is preferred unless there is no choice.
}
\examples{
\dontrun{
start_remote_driver(port = 4567L, browserName = "chrome")
}

}
