% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what_access.R
\name{what_access}
\alias{what_access}
\title{Check paper access type}
\usage{
what_access(articles)
}
\arguments{
\item{articles}{\code{character() | xml_document()}\cr
DOIs of the papers; or the HTML sites containing the papers acquired using
\code{\link[rvest:reexports]{rvest::read_html()}} or \code{\link[httr2:resp_body_raw]{httr2::resp_body_html()}}.}
}
\value{
Paper access type, one of \verb{Open Access}, \verb{Open Archive},
\verb{Free Access}, and \verb{Closed Access}.
}
\description{
Checks the access type under which a paper is available. It can be one of:
"Open Access", "Open Archive", "Free Access", "Closed Access".
}
\section{Supported sources}{
\itemize{
\item Academic Press
\item ACS Publications
\item Analytical Science Journals
\item BioMed Central
\item Biophysical Journal
\item Cambridge Core
\item Cell Reports
\item Cell Reports Medicine
\item eLife
\item Elsevier
\item FEBS Press
\item Frontiers
\item Hindawi
\item Journal of Biological Chemistry
\item Journal of Lipid Research
\item Journal of Neuroscience
\item MDPI
\item Molecular & Cellular Proteomics
\item Molecular Cell
\item Nature
\item OUP Academic
\item PeerJ
\item Pergamon
\item PLOS ONE
\item PNAS
\item Portland Press
\item The Royal Society of Chemistry
\item Science
\item SpringerLink
\item Taylor & Francis
\item Wiley Online Library
}
}

\examples{
what_access("10.1093/nar/gkac882")

}
