% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_open_access.R
\name{is_open_access}
\alias{is_open_access}
\alias{is_open_access.character}
\alias{is_open_access.xml_document}
\alias{is_open_access.oa_html_document}
\title{Check a paper for open access}
\usage{
is_open_access(article)

\method{is_open_access}{character}(article)

\method{is_open_access}{xml_document}(article)

\method{is_open_access}{oa_html_document}(article)
}
\arguments{
\item{article}{\code{character(1) | xml_document(1)}\cr
DOI of the paper; or the HTML site containing the paper acquired using
\code{\link[rvest:reexports]{rvest::read_html()}} or \code{\link[httr2:resp_body_raw]{httr2::resp_body_html()}}.}
}
\value{
\code{TRUE} or \code{FALSE} depending on check result.
}
\description{
Checks whether the full text of the paper is available without restrictions.
Throws error if no interpreter available for the journal.
}
\section{Supported sources}{

\itemize{
\item Academic Press
\item Biophysical Journal
\item eLife
\item Elsevier
\item Frontiers
\item Journal of Biological Chemistry
\item Journal of Neuroscience
\item MDPI
\item Nature
\item PLOS ONE
\item PNAS
\item Portland Press
\item The Royal Society of Chemistry
\item SpringerLink
}
}

\examples{
is_open_access("10.1093/nar/gkac882")

}
